+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+### Fixed
+- resource pullzone: change default for websockets_enabled;
+
 ## 0.11.3 - 2025-11-24
 
 ### Added